requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
drf-spectacular>=0.28,<1.0
psycopg[binary]>=3.2,<4.0
pytest>=8.0,<9.0
pytest-django>=4.5,<5.0
dj-database-url>=2.0,<3.0
python-dotenv>=1.0,<2.0
python-json-logger>=4.0,<5.0
django-redis>=5.4,<6.0
redis>=5.0,<6.0
django-tinymce>=4.0,<5.0
